Add GET endpoint for Jira search
Implemented a new GET endpoint in the router to support retrieving Jira issues based on the provided JQL query and query parameters.
This new endpoint complements the existing POST endpoint in the open-source plugin, providing users with flexibility in choosing the HTTP method for interacting with the plugin.
this enhancement allows users to utilize the response from the GET endpoint for other plugins interacting with Jira.

Signed-off-by: enaysaa saachi.nayyer@ericsson.com