Qoute project key to avoid using JQL reserved keywords
For some reason Jira allows you to create projects that are also JQL keywords. 

Even when using the project ID #, the API will end up using the key. 

The result is that the response contains an error: 

```
"Error in the JQL Query: 'DO' is a reserved JQL word. You must surround it in quotation marks to use it in a query. (line 1, character 13)"
```

Quoting the project key on the JQL should have no impacts and ensure that the proper query is done.